Define renderBlocks
