+++ src/pages/blog/[slug].tsx
@@ -1,16 +1,12 @@
 import React, { useEffect } from 'react'
 import Link from 'next/link'
-import Image from 'next/image'
-import fetch from 'node-fetch'
 import { useRouter } from 'next/router'
 
 import Header from '../../components/header'
-import Heading from '../../components/heading'
 import SocialButtons from '../../components/social-buttons'
-import components from '../../components/dynamic'
 import blogStyles from '../../styles/blog.module.css'
 import { getBlogLink, getTagLink } from '../../lib/blog-helpers'
-import { textBlock } from '../../lib/notion/renderers'
+import { renderBlocks } from '../../lib/notion/renderers'
 import {
   getPosts,
   getAllPosts,
@@ -82,17 +78,6 @@ const RenderPost = ({
 }) => {
   const router = useRouter()
 
-  let listTagName: string | null = null
-  let listLastId: string | null = null
-  let listMap: {
-    [id: string]: {
-      key: string
-      isNested?: boolean
-      nested: string[]
-      children: React.ReactFragment
-    }
-  } = {}
-
   useEffect(() => {
     if (redirect && !post) {
       router.replace(redirect)
@@ -138,140 +123,7 @@ const RenderPost = ({
           </div>
 
           {blocks.length === 0 && <p>This post has no content</p>}
-
-          {blocks.map((block, blockIdx) => {
-            const isLast = blockIdx === blocks.length - 1
-            const isList =
-              block.Type === 'bulleted_list_item' ||
-              block.Type === 'numbered_list_item'
-            let toRender = []
-            let richText
-
-            if (!!block.RichTexts && block.RichTexts.length > 0) {
-              richText = block.RichTexts[0]
-            }
-
-            if (isList) {
-              listTagName =
-                components[block.Type === 'bulleted_list_item' ? 'ul' : 'ol']
-              listLastId = `list${block.Id}`
-
-              listMap[block.Id] = {
-                key: block.Id,
-                nested: [],
-                children: textBlock(block, true, block.Id),
-              }
-            }
-
-            if (listTagName && (isLast || !isList)) {
-              toRender.push(
-                React.createElement(
-                  listTagName,
-                  { key: listLastId! },
-                  Object.keys(listMap).map(itemId => {
-                    if (listMap[itemId].isNested) return null
-
-                    const createEl = item =>
-                      React.createElement(
-                        components.li || 'ul',
-                        { key: item.key },
-                        item.children,
-                        item.nested.length > 0
-                          ? React.createElement(
-                              components.ul || 'ul',
-                              { key: item + 'sub-list' },
-                              item.nested.map(nestedId =>
-                                createEl(listMap[nestedId])
-                              )
-                            )
-                          : null
-                      )
-                    return createEl(listMap[itemId])
-                  })
-                )
-              )
-              listMap = {}
-              listLastId = null
-              listTagName = null
-            }
-
-            const renderHeading = (Type: string | React.ComponentType) => {
-              if (!!richText) {
-                toRender.push(
-                  <Heading key={block.Id}>
-                    <Type key={block.Id}>
-                      {textBlock(block, true, block.Id)}
-                    </Type>
-                  </Heading>
-                )
-              }
-            }
-
-            switch (block.Type) {
-              case 'paragraph':
-                toRender.push(textBlock(block, false, block.Id))
-                break
-              case 'heading_1':
-                renderHeading('h1')
-                break
-              case 'heading_2':
-                renderHeading('h2')
-                break
-              case 'heading_3':
-                renderHeading('h3')
-                break
-              case 'image':
-                toRender.push(
-                  <img src={block.Image.File.Url} alt="image in the content" />
-                )
-                if (
-                  block.Image.Caption.length > 0 &&
-                  block.Image.Caption[0].Text.Content
-                ) {
-                  toRender.push(
-                    <div className={blogStyles.caption}>
-                      {block.Image.Caption[0].Text.Content}
-                    </div>
-                  )
-                }
-                break
-              case 'code':
-                toRender.push(
-                  <components.Code
-                    key={block.Id}
-                    language={block.Language || ''}
-                  >
-                    {block.Code.Text.map(
-                      richText => richText.Text.Content
-                    ).join('')}
-                  </components.Code>
-                )
-                break
-              case 'quote':
-                toRender.push(
-                  React.createElement(
-                    components.blockquote,
-                    { key: block.Id },
-                    block.Quote.Text.map(
-                      richText => richText.Text.Content
-                    ).join('')
-                  )
-                )
-                break
-              default:
-                if (
-                  process.env.NODE_ENV !== 'production' &&
-                  !(
-                    block.Type === 'bulleted_list_item' ||
-                    block.Type === 'numbered_list_item'
-                  )
-                ) {
-                  console.log('unknown type', block.Type)
-                }
-                break
-            }
-            return toRender
-          })}
+          {renderBlocks(blocks)}
           <div>
             <SocialButtons
               title={post.Title}
